added recipe serializer

+++ app/core/models.py
@@ -61,7 +61,7 @@ class Recipe(models.Model):
         on_delete=models.CASCADE
     )
     title = models.CharField(max_length=255)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=5, decimal_places=2)
     time_minutes = models.IntegerField()
     link = models.CharField(max_length=255, blank=True)
     ingredients = models.ManyToManyField('Ingredient')
